MIPS: R5900: Add MFSA and MTSA instructions for the special SA register
The shift amount (SA) register is a 64-bit special register storing the
funnel shift amount. The SA is encoded an implementation-defined manner.
It is therefore not meaningful for software to operate on this value.
Use the MTSAB and MTSAH instructions to set a new funnel shift amount.

The SA is used by the QFSRV (quadword funnel shift right variable)
256-bit multimedia instruction.

MFSA copies the SA register to a 64-bit GPR[1]. The sole purpose of this
instruction is to permit the shift amount to be saved during a context
switch.

MTSA copies a 64-bit GPR rs to the SA register[2]. Note that rs must
contain a value that was originally generated by MFSA. If some other
user-generated value is in rs, the shifting action performed by the
funnel shifter is not defined; that is, MTSA cannot be used to by a
program to set a new funnel shift amount. The sole purpose of this
instruction is to permit the shift amount to be restored during a
context switch. Restrictions:

The three instructions statically preceding a MTSA instruction must
not read or write the SA register; that is, they cannot be either of
the instructions MFSA, QFSRV, or MTSAx.

References:

[1] "TX System RISC TX79 Core Architecture" manual, revision 2.0,
    Toshiba Corporation, p. B-17, https://wiki.qemu.org/File:C790.pdf

[2] Ibid. p. B-20.

Signed-off-by: Fredrik Noring <noring@nocrew.org>