MIPS: PS2: SCMD: Set system command for the real-time clock (RTC)
The hardware clock is designed to keep Japan standard time (JST),
regardless of the region of the machine. This is adjusted in the driver
so that the clock to the kernel appears to be kept in coordinated
universal time (UTC). Tools such as hwclock should therefore set the
clock in the UTC timescale.

Signed-off-by: Fredrik Noring <noring@nocrew.org>